I am going to add my tax calculation algorithm file.
This code file is a model T-Shirt store purchase in which i did a simulation for taxes. I simulated the incidence of ICMS (Brazilian tax paid on goods sold) and PIS/Cofins (Brazilian tax paid on revenue).

Since there has been a recent change on the Supreme Court's understanding of the PIS/Cofins calculation method, I showed both the new way and the old way. The new way does not include ICMS on the calculation basis, while the old one does (quite unfairly).